Refactor TestExecutionManager into focused partial-class files
Splits the 715-line TestExecutionManager.cs into focused partial-class files per issue #9235:

- TestExecutionManager.cs: orchestration entry points (ctors, RunTestsAsync, GetUnitTestDiscoverer)

- TestExecutionManager.Parallelization.cs: ExecuteTestsAsync, ExecuteTestsInSourceAsync

- TestExecutionManager.Runner.cs: ExecuteTestsWithTestRunnerAsync, SendTestResults, MatchTestFilter

- TestExecutionManager.TestContext.cs: GetTestContextProperties, ValidateAndAssignTestProperty, CacheSessionParameters

- TestExecutionManager.Utilities.cs: InitializeRandomTestOrder, Shuffle, RemotingMessageLogger

No behavior or API changes (class remains internal partial). Fixes #9235.
