+++ src/Adapter/MSTestAdapter.PlatformServices/Execution/TestExecutionManager.cs
@@ -1,17 +1,11 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Extensions;
-using Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Helpers;
-using Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.ObjectModel;
-using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices;
 using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Interface;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-using ExecutionScope = Microsoft.VisualStudio.TestTools.UnitTesting.ExecutionScope;
 
 namespace Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Execution;
 
@@ -20,23 +14,8 @@ namespace Microsoft.VisualStudio.TestPlatform.MSTest.TestAdapter.Execution;
 /// </summary>
 #pragma warning disable CA1852 // Seal internal types - This class is inherited in tests.
 [StackTraceHidden]
-internal class TestExecutionManager
+internal partial class TestExecutionManager
 {
-    private sealed class RemotingMessageLogger :
-#if NETFRAMEWORK
-        MarshalByRefObject,
-#endif
-        IMessageLogger
-    {
-        private readonly IMessageLogger _realMessageLogger;
-
-        public RemotingMessageLogger(IMessageLogger messageLogger)
-            => _realMessageLogger = messageLogger;
-
-        public void SendMessage(TestMessageLevel testMessageLevel, string message)
-            => _realMessageLogger.SendMessage(testMessageLevel, message);
-    }
-
     /// <summary>
     /// Dictionary for test run parameters.
     /// </summary>
@@ -196,520 +175,5 @@ internal class TestExecutionManager
 #endif
     }
 
-    /// <summary>
-    /// Execute the parameter tests.
-    /// </summary>
-    /// <param name="tests">Tests to execute.</param>
-    /// <param name="runContext">The run context.</param>
-    /// <param name="frameworkHandle">Handle to record test start/end/results.</param>
-    /// <param name="isDeploymentDone">Indicates if deployment is done.</param>
-    internal virtual async Task ExecuteTestsAsync(IEnumerable<TestCase> tests, IRunContext? runContext, IFrameworkHandle frameworkHandle, bool isDeploymentDone)
-    {
-        InitializeRandomTestOrder(frameworkHandle);
-
-        var testsBySource = (from test in tests
-                             group test by test.Source into testGroup
-                             select new { Source = testGroup.Key, Tests = (IEnumerable<TestCase>)testGroup }).ToArray();
-
-        if (_testOrderRandom is { } random)
-        {
-            Shuffle(random, testsBySource);
-        }
-
-        foreach (var group in testsBySource)
-        {
-            _testRunCancellationToken?.ThrowIfCancellationRequested();
-            await ExecuteTestsInSourceAsync(group.Tests, runContext, frameworkHandle, group.Source, isDeploymentDone).ConfigureAwait(false);
-        }
-    }
-
-    private void InitializeRandomTestOrder(IMessageLogger frameworkHandle)
-    {
-        if (!MSTestSettings.CurrentSettings.RandomizeTestOrder)
-        {
-            _testOrderRandom = null;
-            return;
-        }
-
-        // Use Guid.NewGuid().GetHashCode() rather than new Random() so that consecutive runs do not
-        // collide on the same seed on .NET Framework targets (where Random() is time-seeded with low
-        // resolution). The user can still pin the seed via RandomTestOrderSeed for reproducibility.
-        int seed = MSTestSettings.CurrentSettings.RandomTestOrderSeed ?? Guid.NewGuid().GetHashCode();
-        _testOrderRandom = new Random(seed);
-
-        frameworkHandle.SendMessage(
-            TestMessageLevel.Informational,
-            string.Format(CultureInfo.CurrentCulture, Resource.RandomTestOrderBanner, seed));
-    }
-
-    private static void Shuffle<T>(Random random, T[] items)
-    {
-        // Fisher-Yates shuffle. The array is mutated in place so that callers see a fully
-        // materialized, deterministic order before any parallel work starts.
-        for (int i = items.Length - 1; i > 0; i--)
-        {
-            int j = random.Next(i + 1);
-            (items[i], items[j]) = (items[j], items[i]);
-        }
-    }
-
     internal virtual UnitTestDiscoverer GetUnitTestDiscoverer(ITestSourceHandler testSourceHandler) => new(testSourceHandler);
-
-    internal void SendTestResults(
-        TestCase test,
-        TestTools.UnitTesting.TestResult[] unitTestResults,
-        DateTimeOffset startTime,
-        DateTimeOffset endTime,
-        ITestExecutionRecorder testExecutionRecorder)
-    {
-        foreach (TestTools.UnitTesting.TestResult unitTestResult in unitTestResults)
-        {
-            _testRunCancellationToken?.ThrowIfCancellationRequested();
-
-            var testResult = unitTestResult.ToTestResult(
-                test,
-                startTime,
-                endTime,
-                _environment.MachineName,
-                MSTestSettings.CurrentSettings);
-
-            testExecutionRecorder.RecordEnd(test, testResult.Outcome);
-
-            if (testResult.Outcome == TestOutcome.Failed)
-            {
-                if (PlatformServiceProvider.Instance.AdapterTraceLogger.IsInfoEnabled)
-                {
-                    PlatformServiceProvider.Instance.AdapterTraceLogger.Info("MSTestExecutor:Test {0} failed. ErrorMessage:{1}, ErrorStackTrace:{2}.", testResult.TestCase.FullyQualifiedName, testResult.ErrorMessage, testResult.ErrorStackTrace);
-                }
-
-#if !WINDOWS_UWP && !WIN_UI
-                _hasAnyTestFailed = true;
-#endif
-            }
-
-            try
-            {
-                if (testResult.Outcome != TestOutcome.NotFound
-                    || !RuntimeContext.IsHotReloadEnabled)
-                {
-                    testExecutionRecorder.RecordResult(testResult);
-                }
-            }
-            catch (TestCanceledException)
-            {
-                // Ignore this exception
-            }
-        }
-    }
-
-    private static bool MatchTestFilter(ITestCaseFilterExpression? filterExpression, TestCase test, TestMethodFilter testMethodFilter)
-    {
-        if (filterExpression != null
-            && !filterExpression.MatchTestCase(test, p => testMethodFilter.PropertyValueProvider(test, p)))
-        {
-            // Skip test if not fitting filter criteria.
-            return false;
-        }
-
-        return true;
-    }
-
-    /// <summary>
-    /// Execute the parameter tests present in parameter source.
-    /// </summary>
-    /// <param name="tests">Tests to execute.</param>
-    /// <param name="runContext">The run context.</param>
-    /// <param name="frameworkHandle">Handle to record test start/end/results.</param>
-    /// <param name="source">The test container for the tests.</param>
-    /// <param name="isDeploymentDone">Indicates if deployment is done.</param>
-    private async Task ExecuteTestsInSourceAsync(IEnumerable<TestCase> tests, IRunContext? runContext, IFrameworkHandle frameworkHandle, string source, bool isDeploymentDone)
-    {
-        DebugEx.Assert(!StringEx.IsNullOrEmpty(source), "Source cannot be empty");
-
-#if !WINDOWS_UWP && !WIN_UI
-        if (isDeploymentDone)
-        {
-            source = Path.Combine(PlatformServiceProvider.Instance.TestDeployment.GetDeploymentDirectory()!, Path.GetFileName(source));
-        }
-#endif
-
-        using ITestSourceHost isolationHost = PlatformServiceProvider.Instance.CreateTestSourceHost(source, runContext?.RunSettings);
-        bool usesAppDomains = isolationHost is TestSourceHost { UsesAppDomain: true };
-
-        if (PlatformServiceProvider.Instance.AdapterTraceLogger.IsInfoEnabled)
-        {
-            PlatformServiceProvider.Instance.AdapterTraceLogger.Info("Created unit-test runner {0}", source);
-        }
-
-        // Default test set is filtered tests based on user provided filter criteria
-        ITestCaseFilterExpression? filterExpression = _testMethodFilter.GetFilterExpression(runContext, frameworkHandle, out bool filterHasError);
-        if (filterHasError)
-        {
-            // Bail out without processing everything else below.
-            return;
-        }
-
-        // this is done so that appropriate values of test context properties are set at source level
-        // and are merged with session level parameters
-        IDictionary<string, object> sourceLevelParameters = PlatformServiceProvider.Instance.SettingsProvider.GetProperties(source);
-
-        if (_sessionParameters is { Count: > 0 })
-        {
-            sourceLevelParameters = _sessionParameters.ConcatWithOverwrites(sourceLevelParameters);
-        }
-
-        _testRunCancellationToken?.ThrowIfCancellationRequested();
-
-        TestAssemblySettingsProvider? sourceSettingsProvider = null;
-
-        try
-        {
-            sourceSettingsProvider = isolationHost.CreateInstanceForType(
-                typeof(TestAssemblySettingsProvider),
-                null) as TestAssemblySettingsProvider;
-        }
-        catch (Exception ex)
-        {
-            if (PlatformServiceProvider.Instance.AdapterTraceLogger.IsInfoEnabled)
-            {
-                PlatformServiceProvider.Instance.AdapterTraceLogger.Info("Could not create TestAssemblySettingsProvider instance in child app-domain", ex);
-            }
-        }
-
-        TestAssemblySettings sourceSettings = (sourceSettingsProvider != null)
-            ? sourceSettingsProvider.GetSettings(source)
-            : new TestAssemblySettings();
-
-        int parallelWorkers = sourceSettings.Workers;
-        ExecutionScope parallelScope = sourceSettings.Scope;
-        TestCase[] testsToRun = [.. tests.Where(t => MatchTestFilter(filterExpression, t, _testMethodFilter))];
-        if (_testOrderRandom is { } sourceRandom)
-        {
-            Shuffle(sourceRandom, testsToRun);
-        }
-
-        UnitTestElement[] unitTestElements = [.. testsToRun.Select(e => e.ToUnitTestElementWithUpdatedSource(source))];
-        // Create an instance of a type defined in adapter so that adapter gets loaded in the child app domain
-        var testRunner = (UnitTestRunner)isolationHost.CreateInstanceForType(
-            typeof(UnitTestRunner),
-            [MSTestSettings.CurrentSettings, unitTestElements])!;
-
-        // Ensures that the cancellation token gets through AppDomain boundary.
-        _testRunCancellationToken?.Register(static state => ((UnitTestRunner)state!).Cancel(), testRunner);
-
-        if (MSTestSettings.CurrentSettings.ParallelizationWorkers.HasValue)
-        {
-            // The runsettings value takes precedence over an assembly level setting. Reset the level.
-            parallelWorkers = MSTestSettings.CurrentSettings.ParallelizationWorkers.Value;
-        }
-
-        if (MSTestSettings.CurrentSettings.ParallelizationScope.HasValue)
-        {
-            // The runsettings value takes precedence over an assembly level setting. Reset the level.
-            parallelScope = MSTestSettings.CurrentSettings.ParallelizationScope.Value;
-        }
-
-        if (!MSTestSettings.CurrentSettings.DisableParallelization && sourceSettings.CanParallelizeAssembly && parallelWorkers > 0)
-        {
-            // Parallelization is enabled. Let's do further classification for sets.
-            frameworkHandle.SendMessage(
-                TestMessageLevel.Informational,
-                string.Format(CultureInfo.CurrentCulture, Resource.TestParallelizationBanner, source, parallelWorkers, parallelScope));
-
-            // Create test sets for execution, we can execute them in parallel based on parallel settings
-            // Parallel and not parallel sets.
-            IEnumerable<IGrouping<bool, TestCase>> testSets = testsToRun.GroupBy(t => t.GetPropertyValue(EngineConstants.DoNotParallelizeProperty, false));
-
-            IGrouping<bool, TestCase>? parallelizableTestSet = testSets.FirstOrDefault(g => !g.Key);
-            IGrouping<bool, TestCase>? nonParallelizableTestSet = testSets.FirstOrDefault(g => g.Key);
-
-            if (parallelizableTestSet != null)
-            {
-                ConcurrentQueue<IEnumerable<TestCase>>? queue = null;
-
-                // Chunk the sets into further groups based on parallel level
-                switch (parallelScope)
-                {
-                    case ExecutionScope.MethodLevel:
-                        if (_testOrderRandom is { } methodRandom)
-                        {
-                            IEnumerable<TestCase>[] methodChunks = [.. parallelizableTestSet.Select(t => (IEnumerable<TestCase>)[t])];
-                            Shuffle(methodRandom, methodChunks);
-                            queue = new ConcurrentQueue<IEnumerable<TestCase>>(methodChunks);
-                        }
-                        else
-                        {
-                            queue = new ConcurrentQueue<IEnumerable<TestCase>>(parallelizableTestSet.Select(t => new[] { t }));
-                        }
-
-                        break;
-
-                    case ExecutionScope.ClassLevel:
-                        if (_testOrderRandom is { } classRandom)
-                        {
-                            IEnumerable<TestCase>[] classChunks =
-                            [
-                                .. parallelizableTestSet
-                                    .GroupBy(t => t.GetPropertyValue(EngineConstants.TestClassNameProperty) as string)
-                                    .Select(g => (IEnumerable<TestCase>)g.ToArray()),
-                            ];
-                            Shuffle(classRandom, classChunks);
-                            queue = new ConcurrentQueue<IEnumerable<TestCase>>(classChunks);
-                        }
-                        else
-                        {
-                            queue = new ConcurrentQueue<IEnumerable<TestCase>>(parallelizableTestSet.GroupBy(t => t.GetPropertyValue(EngineConstants.TestClassNameProperty) as string));
-                        }
-
-                        break;
-                }
-
-                var tasks = new List<Task>();
-
-                for (int i = 0; i < parallelWorkers; i++)
-                {
-                    _testRunCancellationToken?.ThrowIfCancellationRequested();
-
-                    tasks.Add(_taskFactory(async () =>
-                    {
-                        try
-                        {
-                            while (!queue!.IsEmpty)
-                            {
-                                _testRunCancellationToken?.ThrowIfCancellationRequested();
-
-                                if (queue.TryDequeue(out IEnumerable<TestCase>? testSet))
-                                {
-                                    await ExecuteTestsWithTestRunnerAsync(testSet, frameworkHandle, source, sourceLevelParameters, testRunner, usesAppDomains).ConfigureAwait(false);
-                                }
-                            }
-                        }
-                        catch (OperationCanceledException) when (_testRunCancellationToken?.Canceled == true)
-                        {
-                        }
-                    }));
-                }
-
-                try
-                {
-                    await Task.WhenAll(tasks).ConfigureAwait(false);
-                }
-                catch (Exception ex)
-                {
-                    string exceptionToString = ex.ToString();
-                    if (PlatformServiceProvider.Instance.AdapterTraceLogger.IsErrorEnabled)
-                    {
-                        PlatformServiceProvider.Instance.AdapterTraceLogger.Error("Error occurred while executing tests in parallel{0}{1}", Environment.NewLine, exceptionToString);
-                    }
-
-                    frameworkHandle.SendMessage(TestMessageLevel.Error, exceptionToString);
-                    throw;
-                }
-            }
-
-            // Queue the non parallel set
-            if (nonParallelizableTestSet != null)
-            {
-                await ExecuteTestsWithTestRunnerAsync(nonParallelizableTestSet, frameworkHandle, source, sourceLevelParameters, testRunner, usesAppDomains).ConfigureAwait(false);
-            }
-        }
-        else
-        {
-            await ExecuteTestsWithTestRunnerAsync(testsToRun, frameworkHandle, source, sourceLevelParameters, testRunner, usesAppDomains).ConfigureAwait(false);
-        }
-
-        if (PlatformServiceProvider.Instance.IsGracefulStopRequested)
-        {
-            testRunner.ForceCleanup(sourceLevelParameters!, new RemotingMessageLogger(frameworkHandle));
-        }
-
-        if (PlatformServiceProvider.Instance.AdapterTraceLogger.IsInfoEnabled)
-        {
-            PlatformServiceProvider.Instance.AdapterTraceLogger.Info("Executed tests belonging to source {0}", source);
-        }
-    }
-
-    private async Task ExecuteTestsWithTestRunnerAsync(
-        IEnumerable<TestCase> tests,
-        ITestExecutionRecorder testExecutionRecorder,
-        string source,
-        IDictionary<string, object> sourceLevelParameters,
-        UnitTestRunner testRunner,
-        bool usesAppDomains)
-    {
-        IEnumerable<TestCase> orderedTests = MSTestSettings.CurrentSettings.OrderTestsByNameInClass && !MSTestSettings.CurrentSettings.RandomizeTestOrder
-            ? tests.OrderBy(t => t.GetManagedType()).ThenBy(t => t.GetManagedMethod())
-            : tests;
-
-        // If testRunner is in a different AppDomain, we cannot pass the testExecutionRecorder directly.
-        // Instead, we pass a proxy (remoting object) that is marshallable by ref.
-        IMessageLogger remotingMessageLogger = usesAppDomains
-            ? new RemotingMessageLogger(testExecutionRecorder)
-            : testExecutionRecorder;
-
-        foreach (TestCase currentTest in orderedTests)
-        {
-            _testRunCancellationToken?.ThrowIfCancellationRequested();
-            if (PlatformServiceProvider.Instance.IsGracefulStopRequested)
-            {
-                break;
-            }
-
-            UnitTestElement unitTestElement = currentTest.ToUnitTestElementWithUpdatedSource(source);
-
-            testExecutionRecorder.RecordStart(currentTest);
-
-            DateTimeOffset startTime = DateTimeOffset.Now;
-
-            if (PlatformServiceProvider.Instance.AdapterTraceLogger.IsInfoEnabled)
-            {
-                PlatformServiceProvider.Instance.AdapterTraceLogger.Info("Executing test {0}", unitTestElement.TestMethod.Name);
-            }
-
-            // Run single test passing test context properties to it.
-            IDictionary<TestProperty, object?>? tcmProperties = TcmTestPropertiesProvider.GetTcmProperties(currentTest);
-            Dictionary<string, object?> testContextProperties = GetTestContextProperties(tcmProperties, sourceLevelParameters, unitTestElement);
-
-            TestTools.UnitTesting.TestResult[] unitTestResult;
-            if (usesAppDomains || Thread.CurrentThread.GetApartmentState() == ApartmentState.STA)
-            {
-#pragma warning disable VSTHRD103 // Call async methods when in an async method - We cannot do right now because we are crossing app domains.
-                // TODO: When app domains support is dropped, we can finally always be calling the async version.
-                // In addition to app domains, if we are STA thread (e.g, because runsettings setting ExecutionApartmentState to STA), we want to preserve that.
-                // If we await, we could end up in a thread pool thread, which is not what we want.
-                // Alternatively, if we want to use RunSingleTestAsync for the case of STA, we should have:
-                // 1. A custom single threaded synchronization context that keeps us in STA.
-                // 2. Use ConfigureAwait(true).
-                unitTestResult = testRunner.RunSingleTest(unitTestElement, testContextProperties, remotingMessageLogger);
-#pragma warning restore VSTHRD103 // Call async methods when in an async method
-            }
-            else
-            {
-                unitTestResult = await testRunner.RunSingleTestAsync(unitTestElement, testContextProperties, remotingMessageLogger).ConfigureAwait(false);
-            }
-
-            if (PlatformServiceProvider.Instance.AdapterTraceLogger.IsInfoEnabled)
-            {
-                PlatformServiceProvider.Instance.AdapterTraceLogger.Info("Executed test {0}", unitTestElement.TestMethod.Name);
-            }
-
-            DateTimeOffset endTime = DateTimeOffset.Now;
-
-            SendTestResults(currentTest, unitTestResult, startTime, endTime, testExecutionRecorder);
-        }
-    }
-
-    /// <summary>
-    /// Get test context properties.
-    /// </summary>
-    /// <param name="tcmProperties">Tcm properties.</param>
-    /// <param name="sourceLevelParameters">Source level parameters.</param>
-    /// <param name="unitTestElement">The unit test element to get properties from.</param>
-    /// <returns>Test context properties.</returns>
-    private static Dictionary<string, object?> GetTestContextProperties(
-        IDictionary<TestProperty, object?>? tcmProperties,
-        IDictionary<string, object> sourceLevelParameters,
-        UnitTestElement unitTestElement)
-    {
-        // If we only have sourceLevelParameters, we create a new dictionary with just those.
-        if (tcmProperties is null &&
-            unitTestElement.Traits is null or { Length: 0 } &&
-            unitTestElement.TestCategory is null or { Length: 0 })
-        {
-            return [with(sourceLevelParameters!)];
-        }
-
-        // To avoid any resizes and additional overhead, we calculate the capacity beforehand.
-        var testContextProperties = new Dictionary<string, object?>(capacity: sourceLevelParameters.Count + (tcmProperties?.Count ?? 0) + (unitTestElement.Traits?.Length ?? 0) + (unitTestElement.TestCategory?.Length ?? 0));
-
-        // Add tcm properties.
-        if (tcmProperties is not null)
-        {
-            foreach (KeyValuePair<TestProperty, object?> kvp in tcmProperties)
-            {
-                testContextProperties[kvp.Key.Id] = kvp.Value;
-            }
-        }
-
-        // Add source level parameters.
-        foreach (KeyValuePair<string, object> kvp in sourceLevelParameters)
-        {
-            testContextProperties[kvp.Key] = kvp.Value;
-        }
-
-        if (unitTestElement.Traits is { Length: > 0 })
-        {
-            foreach (Trait trait in unitTestElement.Traits)
-            {
-                ValidateAndAssignTestProperty(testContextProperties, trait.Name, trait.Value);
-            }
-        }
-
-        if (unitTestElement.TestCategory is { Length: > 0 })
-        {
-            foreach (string category in unitTestElement.TestCategory)
-            {
-                ValidateAndAssignTestProperty(testContextProperties, category, string.Empty);
-            }
-        }
-
-        return testContextProperties;
-    }
-
-    /// <summary>
-    /// Validates If a Custom test property is valid and then adds it to the TestContext property list.
-    /// </summary>
-    /// <param name="testContextProperties"> The test context properties. </param>
-    /// <param name="propertyName"> The property name. </param>
-    /// <param name="propertyValue"> The property value. </param>
-    private static void ValidateAndAssignTestProperty(
-        Dictionary<string, object?> testContextProperties,
-        string propertyName,
-        string propertyValue)
-    {
-        if (StringEx.IsNullOrEmpty(propertyName))
-        {
-            return;
-        }
-
-        if (testContextProperties.ContainsKey(propertyName))
-        {
-            // Do not add to the test context because it would conflict with an already existing value.
-            // We were at one point reporting a warning here. However with extensibility centered around TestProperty where
-            // users can have multiple WorkItemAttributes(say) we cannot throw a warning here. Users would have multiple of these attributes
-            // so that it shows up in reporting rather than seeing them in TestContext properties.
-        }
-        else
-        {
-            testContextProperties.Add(propertyName, propertyValue);
-        }
-    }
-
-    [SuppressMessage("Microsoft.Design", "CA1031:DoNotCatchGeneralExceptionTypes", Justification = "Requirement is to handle errors in user specified run parameters")]
-    private void CacheSessionParameters(IRunContext? runContext, ITestExecutionRecorder testExecutionRecorder)
-    {
-        if (StringEx.IsNullOrEmpty(runContext?.RunSettings?.SettingsXml))
-        {
-            return;
-        }
-
-        try
-        {
-            Dictionary<string, object>? testRunParameters = RunSettingsUtilities.GetTestRunParameters(runContext.RunSettings.SettingsXml);
-            if (testRunParameters != null)
-            {
-                // Clear sessionParameters to prevent key collisions of test run parameters in case
-                // "Keep Test Execution Engine Alive" is selected in VS.
-                _sessionParameters.Clear();
-                foreach (KeyValuePair<string, object> kvp in testRunParameters)
-                {
-                    _sessionParameters.Add(kvp);
-                }
-            }
-        }
-        catch (Exception ex)
-        {
-            testExecutionRecorder.SendMessage(TestMessageLevel.Error, ex.Message);
-        }
-    }
 }
